Update eslintrc
Turned on warnings for console-logs since this is a browser project and
console logs would only be used for debugging. They should be removed
for production unless the product needs it for some reason.

Added import/extensions rule to throw an error when there is no ending
extension for local imports. This makes it clear the imported file is
a js file or something else. It ignores it for packages because those
could be anything (js, mjs, cjs), so it leaves it to the resolver to
decide.